Revert "Use long max_attempts to avoid casting, feed directly to AWS API"
This reverts commit 7fcbed119e5876f89b6bace77be743e50c8d7803.

+++ cpp/src/arrow/filesystem/s3fs.cc
@@ -243,15 +243,17 @@ class AwsRetryStrategy : public S3RetryStrategy {
 };
 
 std::shared_ptr<S3RetryStrategy> S3RetryStrategy::GetAwsDefaultRetryStrategy(
-    long max_attempts) {
+    int64_t max_attempts) {
   return std::make_shared<AwsRetryStrategy>(
-      std::make_shared<Aws::Client::DefaultRetryStrategy>(max_attempts));
+      std::make_shared<Aws::Client::DefaultRetryStrategy>(
+          static_cast<long>(max_attempts)));
 }
 
 std::shared_ptr<S3RetryStrategy> S3RetryStrategy::GetAwsStandardRetryStrategy(
-    long max_attempts) {
+    int64_t max_attempts) {
   return std::make_shared<AwsRetryStrategy>(
-      std::make_shared<Aws::Client::StandardRetryStrategy>(max_attempts));
+      std::make_shared<Aws::Client::StandardRetryStrategy>(
+          static_cast<long>(max_attempts)));
 }
 
 // -----------------------------------------------------------------------
